+++ Sources/Base/OAuth2AuthConfig.swift
@@ -18,14 +18,14 @@
 //  limitations under the License.
 //
 
-
-/**
-Simple struct to hold settings describing how authorization appears to the user.
-*/
 #if os(iOS)
 import UIKit
 #endif
 
+
+/**
+Simple struct to hold settings describing how authorization appears to the user.
+*/
 public struct OAuth2AuthConfig {
 	
 	/// Sub-stuct holding configuration relevant to UI presentation.