+++ Sources/Base/OAuth2AuthRequest.swift
@@ -80,6 +80,7 @@ open class OAuth2AuthRequest {
 	/// Custom headers can be set here, they will take precedence over any built-in headers.
 	open private(set) var headers: [String: String]?
 	
+	/// Query parameters to use with the request.
 	open var params = OAuth2RequestParams()